tests: Add firmware-auto-efi-format-nvram-raw-nvramtemplate-path
This test case demonstrates an issue with the current
implementation of firmware autoselection.

There is no requirement for the format of the NVRAM file (raw in
this case) to match that of the NVRAM template (qcow2 in this
case), and yet libvirt incorrectly rejects the configuration.

The issue will be addressed in an upcoming commit.

Signed-off-by: Andrea Bolognani <abologna@redhat.com>
Reviewed-by: Michal Privoznik <mprivozn@redhat.com>